refactor: rename scene engine shared to core
