+++ frontend/src/features/scene-engine/core/scene-engine.constants.ts
