+++ frontend/src/features/scene-engine/core/scene-engine.queries.ts
