+++ frontend/src/features/scene-engine/core/scene-engine.types.ts
