+++ frontend/src/features/scene-engine/steps/character/character.types.ts
@@ -1,4 +1,4 @@
-import type { ImageRecord } from "@scene-engine/shared/scene-engine.types";
+import type { ImageRecord } from "@scene-engine/core/scene-engine.types";
 
 export type CharacterRecord = {
   id: string;
