+++ frontend/src/features/scene-engine/steps/character/character.utils.ts
@@ -1,6 +1,6 @@
 import { isAxiosError } from "axios";
 import { httpClient } from "@/lib/httpClient";
-import { STORY_API_BASE } from "@scene-engine/shared/scene-engine.constants";
+import { STORY_API_BASE } from "@scene-engine/core/scene-engine.constants";
 import type { CharacterBundle } from "./character.types";
 
 export function spliceCharacterIntoList(
