+++ frontend/src/features/scene-engine/steps/panel/extraction/components/RefImageTray.tsx
@@ -1,7 +1,7 @@
 import { useQuery, useQueryClient } from "@tanstack/react-query";
 import { Skeleton } from "@scene-engine/components/Skeleton";
-import { imageSignedUrlOptions } from "@scene-engine/shared/scene-engine.queries";
-import type { ImageRecord } from "@scene-engine/shared/scene-engine.types";
+import { imageSignedUrlOptions } from "@scene-engine/core/scene-engine.queries";
+import type { ImageRecord } from "@scene-engine/core/scene-engine.types";
 
 function RefImageThumbnail({
   img,
