+++ frontend/src/features/scene-engine/steps/panel/extraction/hooks/usePanelExtraction.ts
@@ -1,5 +1,5 @@
 import { httpClient } from "@/lib/httpClient";
-import { STORY_API_BASE } from "@scene-engine/shared/scene-engine.constants";
+import { STORY_API_BASE } from "@scene-engine/core/scene-engine.constants";
 import { useMutation, useQuery, useQueryClient } from "@tanstack/react-query";
 import { panelsOptions } from "../../panel.queries";
 import { splicePanelIntoList, uploadReferenceImageRequest, buildHttpErrorMessage } from "../../panel.utils";
