+++ frontend/src/features/scene-engine/steps/panel/panel.queries.ts
@@ -1,6 +1,6 @@
 import { httpClient } from "@/lib/httpClient";
 import { queryOptions } from "@tanstack/react-query";
-import { STORY_API_BASE, STORY_QUERY_ROOT } from "@scene-engine/shared/scene-engine.constants";
+import { STORY_API_BASE, STORY_QUERY_ROOT } from "@scene-engine/core/scene-engine.constants";
 import type { PanelBundle } from "./panel.types";
 
 export const panelsOptions = (projectId: string, storyId: string) =>
