+++ frontend/src/features/scene-engine/steps/panel/panel.utils.ts
@@ -1,6 +1,6 @@
 import { isAxiosError } from "axios";
 import { httpClient } from "@/lib/httpClient";
-import { STORY_API_BASE } from "@scene-engine/shared/scene-engine.constants";
+import { STORY_API_BASE } from "@scene-engine/core/scene-engine.constants";
 import type { PanelBundle } from "./panel.types";
 
 export function splicePanelIntoList(
