+++ frontend/src/features/scene-engine/steps/review/components/ComicCarousel.tsx
@@ -1,9 +1,9 @@
 import { useEffect, useRef } from "react";
 import { useQuery, useQueryClient } from "@tanstack/react-query";
 import { Skeleton } from "@scene-engine/components/Skeleton";
-import { imageSignedUrlOptions } from "@scene-engine/shared/scene-engine.queries";
+import { imageSignedUrlOptions } from "@scene-engine/core/scene-engine.queries";
 import type { PanelBundle } from "../../panel/panel.types";
-import type { ImageRecord } from "@scene-engine/shared/scene-engine.types";
+import type { ImageRecord } from "@scene-engine/core/scene-engine.types";
 
 function PanelImage({ render }: { render: ImageRecord }) {
   const queryClient = useQueryClient();
